devops/jenkins/agent/requirements.txt
python-jenkins >= 1.7.0
requests >= 2.25.1
attrs >= 20.3.0
docker >= 4.0.2
